requirements.txt
APScheduler==3.10.4
alembic==1.13.1
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.135.3
GeoAlchemy2==0.14.3
greenlet==3.3.2
gunicorn==21.2.0
h11==0.16.0
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
pytest==9.0.3
requests==2.33.1
Shapely==2.0.4
slowapi==0.1.9
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==8.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
